Refactoring Tests

+++ entity-locker-lib/src/test/java/org/locker/tests/GlobalEntityLockerTest.java
@@ -347,7 +347,7 @@ class GlobalEntityLockerTest {
         while (i-- != 0) {
             int numberOfLocks = (int) (random() * 3 + 1);
             for (int j = 0; j < numberOfLocks; j++) {
-                double lockType = random() * 2;
+                int lockType = (int) (random() * 2);
                 if (lockType == 0) {
                     locker.lock();
                 } else {
